modification mineure

+++ views/admin/v_consulter_client.php
@@ -17,8 +17,8 @@
     </head>
     <body>
         <?php
-        include("views/admin/_v_header.php") ;
-        include("views/admin/_v_menu.php") ;
+        include("views/admin/_v_header.php");
+        include("views/admin/_v_menu.php");
         ?>
         <div id="contenu">
             <?php showNotifications() ?>
@@ -26,8 +26,10 @@
             <div>
                 <div id="breadcrumb">
                     <a href="index.php?uc=gererClients&action=listerClients">Retour à la liste</a>&nbsp;
-                    <a href="index.php?uc=gererClients&action=modifierClient&id=<?php echo $intNoClient ?>">Modifier</a>&nbsp;
-                    <a href="index.php?uc=gererClients&action=supprimerClient&id=<?php echo $intNoClient ?>">Supprimer</a>
+                    <?php if ($_SESSION['profil'] == 1) { ?>
+                        <a href="index.php?uc=gererClients&action=modifierClient&id=<?php echo $intNoClient ?>">Modifier</a>&nbsp;
+                        <a href="index.php?uc=gererClients&action=supprimerClient&id=<?php echo $intNoClient ?>">Supprimer</a>
+                    <?php } ?>
                 </div>
                 <table>
                     <tr>
@@ -90,49 +92,47 @@
                 </table>
                 <h3>Contrats</h3>
                 <?php if ($lesContrats) { ?>
-                <div id="objectList">
-                    <table>
-                        <tr>
-                            <th class="id">N°</th>
-                            <th>Date</th>
-                            <th>Céréale</th>
-                            <th>Commandé</th>
-                            <th>Prix</th>
-                            <th>Livré</th>
-                            <th>Etat</th>
-                        </tr>
-                        <?php
-                        $qteTotale = 0;
-                        $i = 0;
-                        foreach ($lesContrats as $unContrat) {
-                            if ($i % 2) {
-                                echo '<tr class="pair">';
+                    <div id="objectList">
+                        <table>
+                            <tr>
+                                <th class="id">N°</th>
+                                <th>Date</th>
+                                <th>Céréale</th>
+                                <th>Commandé</th>
+                                <th>Prix</th>
+                                <th>Livré</th>
+                                <th>Etat</th>
+                            </tr>
+                            <?php
+                            $qteTotale = 0;
+                            $i = 0;
+                            foreach ($lesContrats as $unContrat) {
+                                if ($i % 2) {
+                                    echo '<tr class="pair">';
+                                } else {
+                                    echo '<tr class="impair">';
+                                }
+                                echo '<td class="id"><a href="index.php?page=consulterContrat&id=' . $unContrat[0] . '">' . $unContrat[0] . '</a></td>';
+                                echo '<td>' . $unContrat[1] . '</td>';
+                                echo '<td>' . $unContrat[2] . '</td>';
+                                echo '<td>' . $unContrat[3] . '</td>';
+                                echo '<td>' . $unContrat[4] . '</td>';
+                                echo '<td>' . $unContrat[5] . '</td>';
+                                echo '<td>' . $unContrat[6] . '</td>';
+                                echo '</tr>';
+                                $i++;
+                                $qteTotale += $unContrat[3];
                             }
-                            else {
-                                echo '<tr class="impair">';
-                            }
-                            echo '<td class="id"><a href="index.php?page=consulterContrat&id='.$unContrat[0].'">'.$unContrat[0].'</a></td>';
-                            echo '<td>'.$unContrat[1].'</td>';
-                            echo '<td>'.$unContrat[2].'</td>';
-                            echo '<td>'.$unContrat[3].'</td>';
-                            echo '<td>'.$unContrat[4].'</td>';
-                            echo '<td>'.$unContrat[5].'</td>';
-                            echo '<td>'.$unContrat[6].'</td>';
-                            echo '</tr>';
-                            $i++;
-                            $qteTotale += $unContrat[3];
-                        }
-                        ?>
-                    </table>
-                    <p>Total : <?php echo $qteTotale." t" ?></p>
-                </div>
-            <?php 
-            } 
-            else {
-                echo "<p>Aucun contrat trouvé pour ce client</p>";
-            }
-            ?>                        
-          </div>
+                            ?>
+                        </table>
+                        <p>Total : <?php echo $qteTotale . " t" ?></p>
+                    </div>
+                    <?php
+                } else {
+                    echo "<p>Aucun contrat trouvé pour ce client</p>";
+                }
+                ?>                        
+            </div>
         </div>
     </body>
 </html>
